🛠️ refactor: (FE) - socketEmit을 useSocket 훅으로 변경
